feat(tui): render agent replies as markdown, split tui package
